requirements.txt
# DataHub SDK for emitting metadata and lineage
acryl-datahub

# Firebase Admin SDK for real-time database listeners
firebase-admin